// src/features/sales/msvo/muse-vectan-and-other.dto.ts
export interface MuseVectanAndOtherDTO {
  _id?: string;
  buyer: string;
  country: string;
  cotractNumber: string;
  contractDate: string;
  contractSumm: number;
  actualStock: number;
  invoiceNumber: string;
  lotNumber: string;
  shipmentDate: string;
  quantity: number;
  invoiceSumm: number;
  deliveryStatus: string;
  paymentConditions: string;
  paymentDate: string;
  comment: string;
  createdAt?: string;
  updatedAt?: string;
}